+++ ostap/fitting/simfit.py
@@ -1362,18 +1362,18 @@ class SimFit (VarMaker,ConfigReducer) :
         return self.pdf.__getitem__ ( param )
 
     # =========================================================================
-    ## get NLL/profile-graph for the variable, using the specified bscissas
+    ## get NLL/profile-graph for the variable, using the specified abscissas
     #  @code
     #  pdf   = ...
-    #  graph = pdf.graph_nll ( 'S'               ,
-    #                          range ( 0 , 100 ) ,
-    #                          dataset           )
+    #  graph = pdf.graph_nll ( 'S'                ,
+    #                          vrange ( 0 , 100 ) ,
+    #                          dataset            )
     #  @endcode
     def graph_nll ( self , *args , **kwargs ) :
         """ Get NLL/profile-graph for the variable, using the specified abscissas
         >>> pdf   = ...
         >>> graph = pdf.graph_nll ( 'S'               ,
-        ...                          range ( 0 , 100 ) ,
+        ...                          vrange ( 0 , 100 ) ,
         ...                          dataset           )
         """
         return self.pdf.graph_nll ( *args , **kwargs )
@@ -1389,9 +1389,9 @@ class SimFit (VarMaker,ConfigReducer) :
     def graph_profile ( self , *args , **kwargs ) :
         """ Get profile-graph for the variable, using the specified abscissas
         >>> pdf   = ...
-        >>> graph = pdf.graph_profile ( 'S'                     ,
-        ...                             range ( 0 , 12.5 , 20 ) ,
-        ...                             dataset                 )
+        >>> graph = pdf.graph_profile ( 'S'                      ,
+        ...                             vrange ( 0 , 12.5 , 20 ) ,
+        ...                             dataset                  )
         """
         return self.pdf.graph_profile ( *args , **kwargs )